# tictactoe-game/apidoc.json
{"openapi":"3.0.0","paths":{"/game":{"get":{"operationId":"GameController_getActiveGame","parameters":[{"name":"Authorization","in":"header","description":"JWT Token for Authorization","schema":{"type":"string"}}],"responses":{"200":{"description":"The board state of the current Game","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameState"}}}},"401":{"description":"No valid JWT Token provided"}},"security":[{"bearer":[]}],"tags":["Game"]}},"/game/move/user":{"post":{"operationId":"GameController_moveUser","parameters":[{"name":"Authorization","in":"header","description":"JWT Token for Authorization","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoveDTO"}}}},"responses":{"200":{"description":"The board state after the move","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameState"}}}},"400":{"description":"Invalid Field was selected"},"401":{"description":"No valid JWT Token provided"}},"security":[{"bearer":[]}],"tags":["Game"]}},"/game/move/opponent":{"post":{"operationId":"GameController_moveOpponent","parameters":[{"name":"Authorization","in":"header","description":"JWT Token for Authorization","schema":{"type":"string"}}],"responses":{"200":{"description":"The board state after the move","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameState"}}}},"400":{"description":"Invalid Field was selected"},"401":{"description":"No valid JWT Token provided"}},"security":[{"bearer":[]}],"tags":["Game"]}}},"info":{"title":"TicTacToe Game","description":"Sample Implementation for a TicTacToe Game","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"GameState":{"type":"object","properties":{"id":{"type":"string","description":"ID of the Game"},"playerId":{"type":"string","description":"Current Player's identifier"},"board":{"description":"Board stored as flat array","type":"array","items":{"type":"string","maxLength":9,"minLength":9}},"turn":{"type":"number","description":"Current Turn index","minimum":0,"maximum":8},"finished":{"type":"boolean","description":"Information whether the game is finished"},"winner":{"type":"string","description":"ID of the winner. Null if no winner exists"}},"required":["id","playerId","board","turn","finished","winner"]},"MoveDTO":{"type":"object","properties":{"row":{"type":"number","description":"Row to put the Move","minimum":0,"maximum":2},"column":{"type":"number","description":"Column to put the Move","minimum":0,"maximum":2}},"required":["row","column"]}}}}